+++ app/src/main/java/com/thailam/weatherwhen/data/model/DisplayedForecast.kt
@@ -2,7 +2,7 @@ package com.thailam.weatherwhen.data.model
 
 data class DisplayedForecast(
     val id : Int,
-    val epochDate : Int,
+    val epochDate : Long,
     val temperatureValue : Double,
     val temperatureUnit : String,
     val currentCondition: CurrentCondition
